Added SqurellRandom
Former-commit-id: 46d58c1d7a276199bae237042c3e7066130dc1dc

+++ EchoRenderer/Rendering/Engines/ProgressiveRenderEngine.cs
@@ -170,7 +170,7 @@ namespace EchoRenderer.Rendering.Engines
 			long tick = Environment.TickCount64;
 
 			RenderProfile profile = CurrentProfile;
-			int seed = HashCode.Combine(id, tick);
+			uint seed = (uint)HashCode.Combine(id, tick);
 			return profile.Method.CreateArena(profile, seed);
 		}
